requirements.txt
google-genai>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
lxml>=4.9.0
